--- pkg/app/src/containers/Account/Settings.tsx
@@ -1,12 +1,0 @@
-import React from 'react'
-
-import Box from '@mui/material/Box'
-
-interface ComponentProps {
-	args: AccountTabs
-}
-export function Settings({ args }: ComponentProps) {
-	return <Box></Box>
-}
-
-export default Settings
